+++ src/utils/api/fbInstance.js
@@ -14,4 +14,6 @@ const firebaseConfig = {
   measurementId: 'G-JC54NGCVTD',
 };
 
-export default firebase.initializeApp(firebaseConfig);
+firebase.initializeApp(firebaseConfig);
+
+export default firebase;
